fix some errors in comments causing documentation generation problems
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/trunk@28250 72102866-910b-0410-8b05-ffd578937521

+++ Source/NSDecimal.m
@@ -1120,7 +1120,7 @@ NSSimpleCompare(const NSDecimal *leftOperand, const NSDecimal *rightOperand)
 }
 
 void
-NSDecimalRound(NSDecimal *result, const NSDecimal *number, int scale,
+NSDecimalRound(NSDecimal *result, const NSDecimal *number, NSInteger scale,
 	       NSRoundingMode mode)
 {
   GSDecimal m;
